% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/SpotifyAPI.R
\name{getAlbumsTracks}
\alias{getAlbumsTracks}
\title{Get Album Tracks}
\usage{
getAlbumsTracks(getArtistsAlbumsOutput)
}
\arguments{
\item{getArtistsOutput}{A data.frame object produced from the getArtists function}

\item{country}{the country the albums are availble in by default GB}

\item{albumType}{the album types to select. By defult album. Options: album, single, compilation, appears_on}

\item{cleanDups}{some results contain duplicate albums due to reissues. TRUE by default will clean this up}
}
\description{
Creates a data frame containing details of an artists albums taken from the output from getArtists. This is limited by Spotify to a maximum of 50 albums.
}
\examples{
tracks<-getAlbumsTracks(getArtistsAlbums(getArtists("The Veils")))
}
\author{
James Thomson
}

