% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/SpotifyAPI.R
\name{getArtists}
\alias{getArtists}
\title{Get Artists}
\usage{
getArtists(artists)
}
\arguments{
\item{artists}{A value or vector of artists to search for}
}
\description{
Creates a data frame containing details of the top result for the searched for artists
}
\examples{
result<-getArtists("David Bowie")
results<-getArtists(c("David Bowie", "Bob Dylan"))
}
\author{
James Thomson
}

