% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/SpotifyAPI.R
\name{getRelatedArtists}
\alias{getRelatedArtists}
\title{Get Related Artists}
\usage{
getRelatedArtists(getArtistsOutput)
}
\arguments{
\item{GetArtistsOutput}{A data.frame object produced from the getArtists function}
}
\description{
Creates a data frame containing details of the related artists taken from the output from getArtists
}
\examples{
artist<-getArtists(c("David Bowie", "Bob Dylan"))
related_artist<-getRelatedArtists(artist)
}
\author{
James Thomson
}

