% Generated by roxygen2 (4.0.1): do not edit by hand
\name{visDiscography}
\alias{visDiscography}
\title{Visualize Artists Album History}
\usage{
visDiscography(artist, output_file)
}
\arguments{
\item{artist}{an artist}

\item{output_file}{an output location and filename}
}
\description{
Creates a object containing an artists back catalogue and a visualisation of their back catalogue at the file location specified
}
\examples{
out<-visDiscography(artist="David Bowie", output_file="Disco.html")
}
\author{
James Thomson
}

