% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/SpotifyAPI.R
\name{visRelatedArtists}
\alias{visRelatedArtists}
\title{Visualize Related Artists}
\usage{
visRelatedArtists(artist, steps = 2, output_file)
}
\arguments{
\item{artist}{an artist or list of artists}

\item{steps}{the number of steps you want to explore related artists for}

\item{output_file}{an output location and filename}
}
\description{
Creates a object containing a list of artists and a list of artists relationships then outputs
a force direct layout to the file location specified
}
\examples{
out<-visRelatedArtists(artist="David Bowie", steps=2, output_file="vis.html")
out<-visRelatedArtists(artist=c("David Bowie", "Bob Dylan"), steps=1, output_file="vis.html")
}
\author{
James Thomson
}

